==> Proyecto 1/notebook/Draft/01. Selenium_limpieza_.ipynb <==
import pandas as pd
import numpy as np
import sidetable
import requests
from time import sleep
import re

from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.chrome.options import Options

import warnings
warnings.filterwarnings('ignore')

# CSV Inicial

Nos descargamos el CSV que vamos a utilizar de Kaggle.

Los datos son sobre casas en los Estados de Washington, California y Oregón, y divididos por ciudades. Incluye muchas de las ciudades más importantes de cada estado. Los datos incluyen el precio, dirección, número de cuartos, baños, el área de la casa, etc.

Link al kaggle: https://www.kaggle.com/ialjadani/houses-prices-in-selected-area-in-usa

df = pd.read_csv ("../Data/01. CSV_inicial_houses_USA.csv")
df.sort_values(by = ["City"]).head(3)

df.shape

# Selenium

A continuación, mediante Selenium, vamos a extraer información adicional sobre índices (de poder adquisitivo, calidad, seguridad, sanidad, etc.) de las 11 ciudades más importantes y relevantes de estos 3 Estados. Esto lo usaremos más adelante para enriquecer nuestro CSV inicial

Lo primero que vamos a hacer es definir una serie de opciones para trabajar con Selenium

opciones= Options()
opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
#para ocultarme como robot
opciones.add_experimental_option('useAutomationExtension', False)
opciones.add_argument('--start-maximized') #empezar maximizado
opciones.add_argument('user.data-dir=selenium') #guarda las cookies
opciones.add_argument('--incognito')#incognito window

# iniciamos el driver
driver = webdriver.Chrome(ChromeDriverManager().install())
## accedemos a la pagina web

driver.get("https://es.numbeo.com/calidad-de-vida/pa%25C3%25ADs/Estados-Unidos")
driver.implicitly_wait(15)
# esperamos

driver.find_element_by_css_selector("#content_and_logo > table:nth-child(10) > tbody").text

Hemos accedido a la información genérica de Estados Unidos, perfecto! Y ahora vamos a acceder a la información de las ciudades que nos interesan:

ciudades_list = ["Olympia","Portland","Seattle","Anaheim","Bellevue","Irvine","Playa-Larga","Los-Ángeles","Pasadena","Portland","Redmond","Tacoma","Vancouver"]         

url_list = []

for i in ciudades_list:
    url_list.append(f"https://es.numbeo.com/calidad-de-vida/ciudad/{i}")

url_list

Ya tenemos la lista de las url que queremos! Ahora vamos a sacar la tabla para cada una de las url de arriba:

resultados = []
for z in url_list:
    driver.get(z)
    driver.implicitly_wait(15)
    res = driver.find_element_by_xpath('/html/body/div[2]/table[2]/tbody').text
    resultados.append(res)

resultados

A continuación vamos a crear un diccionario, pasándole los Key-values, y con objetivo de crear un DataFrame:

key_list = ["Olympia","Portland","Seattle","Anaheim","Bellevue","Irvine","Long Beach","Los Ángeles","Pasadena","Portland","Redmond","Tacoma","Vancouver"]
value_list = resultados

dict_from_list = dict(zip(key_list, value_list))
#print(dict_from_list)

df1 = pd.DataFrame(dict_from_list, index = ["indice1"]).T
df1

A continuación vamos a split la columna "indice1" por los espacios! Y resetemos el índice:

df2 = (df1["indice1"].str.split("\n", expand=True)).reset_index()
df2.head(1)

Actualizamos el nombre de las columnas

df2.columns = ['City', 'índice_poder_adq', 'índice_seguridad','índice_sanidad', 'índice_clima', 'índice_costo_vida', 'relación_precio_ingresos_propiedades', 'índice_costo_desplazamiento', 'índice_contaminación', 'índice_calidad_vida']
df2.head(1)

Creamos nuevas columnas, quedándonos únicamente con los valores numéricos:

df2["Índice_poder_adqui"] = df2.índice_poder_adq.str.extract("(\d+,\d+)", expand=True)
df2["Índice_seguridad"] = df2.índice_seguridad.str.extract("(\d+,\d+)", expand=True)
df2["Índice_sanidad"] = df2.índice_sanidad.str.extract("(\d+,\d+)", expand=True)
df2["Índice_clima"] = df2.índice_clima.str.extract("(\d+,\d+)", expand=True)
df2["Índice_costo_vida"] = df2.índice_costo_vida.str.extract("(\d+,\d+)", expand=True)
df2["Relación_precio_ingresos_propiedades"] = df2.relación_precio_ingresos_propiedades.str.extract("(\d+,\d+)", expand=True)
df2["Índice_costo_desplazamiento"] = df2.índice_costo_desplazamiento.str.extract("(\d+,\d+)", expand=True)
df2["Índice_contaminación"] = df2.índice_contaminación.str.extract("(\d+,\d+)", expand=True)
df2["Índice_calidad_vida"] = df2.índice_calidad_vida.str.extract("(\d+,\d+)", expand=True)
df2.head(1)

Eliminamos las columnas que ya no necesitamos porque la información está duplicada:

eliminar = ['índice_poder_adq', 'índice_seguridad', 'índice_sanidad',
       'índice_clima', 'índice_costo_vida',
       'relación_precio_ingresos_propiedades', 'índice_costo_desplazamiento',
       'índice_contaminación', 'índice_calidad_vida']

df2_limpio = df2.drop(eliminar, axis = 1)
df2_limpio.head(1)

Exportamos el DataFrame a un CSV

df2_limpio.to_csv("01. Resultado_Selenium_índices_ciudades_EEUU.csv")
